oddoAG/requirements.txt
flask==3.1.1
flask-cors==5.0.1
sqlalchemy==2.0.40
pyjwt==2.10.1
werkzeug==3.1.3
requests==2.32.3
flask-mail==0.10.0
pytesseract==0.3.13
Pillow==11.1.0
python-dotenv==1.1.0
